requirements.txt
jax[cuda]
jaxlib
flax
optax
numpy
matplotlib
opencv-python
Pillow
fonttools
tqdm